third_party/WebKit/LayoutTests/http/tests/background_fetch/background-fetch-manager-get.https.html

<!doctype html>
<meta charset="utf-8">
<title>Background Fetch API: get() tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/test-helpers.js"></script>

<h1>BackgroundFetchManager.get()</h1>
<p>This test validates the behaviour of the get() method.</p>

<!-- TODO(peter): Move this to the WPT directory when it's merged. -->

<script>
'use strict';

const workerUrl = 'resources/empty-worker.js';
const scope = 'resources/scope/' + location.pathname;

promise_test(function(test) {
  return service_worker_unregister_and_register(test, workerUrl, scope)
    .then(registration => {
      assert_equals(null, registration.active);
      return registration.backgroundFetch.get('id');
    })
    .then(unreached_fulfillment(test), error => {
      assert_equals(error.name, 'TypeError');
    });

}, 'BackgroundFetchManager.get() requires an activated Service Worker.');

promise_test(function(test) {
  let registration = null;

  return service_worker_unregister_and_register(test, workerUrl, scope)
    .then(r => {
      registration = r;
      return wait_for_state(test, r.installing, 'activated');
    })
    .then(() => registration.backgroundFetch.get('invalid-fetch-id'))
    .then(backgroundFetchRegistration => {
      assert_equals(backgroundFetchRegistration, null);
    });

}, 'BackgroundFetchManager.get() returns NULL for non-existing fetches.');

promise_test(function(test) {
  const id = 'my-id';
  const options = {
    icons: [
      {
        src: 'resources/non-existing-large-icon.png',
        sizes: '256x256',
        type: 'image/png'
      },
      {
        src: 'resources/non-existing-small-icon.jpg',
        sizes: '64x64',
        type: 'image/jpg'
      }
    ],
    title: 'My Background Fetch',
    downloadTotal: 1024
  };

  let registration = null;

  return service_worker_unregister_and_register(test, workerUrl, scope)
    .then(r => {
      registration = r;
      return wait_for_state(test, r.installing, 'activated');
    })
    .then(() => registration.backgroundFetch.fetch(id, ['resources/non-existing-file.png'], options))
    .then(backgroundFetchRegistration => {
      assert_true(backgroundFetchRegistration instanceof BackgroundFetchRegistration);

      return registration.backgroundFetch.get(id)
        .then(secondBackgroundFetchRegistration => {
          assert_true(secondBackgroundFetchRegistration instanceof BackgroundFetchRegistration);
          assert_equals(secondBackgroundFetchRegistration.id, id);
          assert_equals(secondBackgroundFetchRegistration.downloadTotal, options.downloadTotal);
        });
    });
}, 'BackgroundFetchManager.get() can return created fetches.');

promise_test(function(test) {
  const id = 'my-id';

  let registration = null;

  return service_worker_unregister_and_register(test, workerUrl, scope)
    .then(r => {
      registration = r;
      return wait_for_state(test, r.installing, 'activated');
    })
    .then(() => registration.backgroundFetch.fetch(id, ['resources/non-existing-file.png']))
    .then(backgroundFetchRegistration => {
      assert_true(backgroundFetchRegistration instanceof BackgroundFetchRegistration);
      return registration.backgroundFetch.get(id);
    })
    .then(secondBackgroundFetchRegistration => {
      assert_true(secondBackgroundFetchRegistration instanceof BackgroundFetchRegistration);
      return secondBackgroundFetchRegistration.abort();
    })
    .then(() => registration.backgroundFetch.get(id))
    .then(thirdBackgroundFetchRegistration => {
      assert_equals(thirdBackgroundFetchRegistration, null);
    });

}, 'BackgroundFetchManager.get() returned fetches can be aborted.');
</script>
