third_party/WebKit/LayoutTests/platform/linux/http/tests/background_fetch/background-fetch-manager-get.https-expected.txt

This is a testharness.js-based test.
PASS BackgroundFetchManager.get() requires an activated Service Worker.
PASS BackgroundFetchManager.get() returns NULL for non-existing fetches.
PASS BackgroundFetchManager.get() can return created fetches.
FAIL BackgroundFetchManager.get() returned fetches can be aborted. assert_equals: expected null but got object "[object BackgroundFetchRegistration]"
Harness: the test ran to completion.
